+++ Source/Dafny/DafnyOptions.cs
@@ -84,6 +84,7 @@ namespace Microsoft.Dafny {
     public string DafnyPrintCompiledFile = null;
     public string CoverageLegendFile = null;
     public string MainMethod = null;
+    public bool RunAllTests = false;
     public bool ForceCompile = false;
     public bool RunAfterCompile = false;
     public int SpillTargetCode = 0; // [0..4]
@@ -262,6 +263,15 @@ namespace Microsoft.Dafny {
         case "main": {
             if (ps.ConfirmArgumentCount(1)) {
               MainMethod = args[ps.i];
+            }
+
+            return true;
+          }
+
+        case "runAllTests": {
+            int runAllTests = 0;
+            if (ps.GetIntArgument(ref runAllTests, 2)) {
+              RunAllTests = runAllTests != 0; // convert to boolean
             }
 
             return true;
@@ -1130,6 +1140,15 @@ Exit code: 0 -- success; 1 -- invalid command-line; 2 -- parse or type errors;
 /Main:<name>
     The (fully-qualified) name of the method to use as the executable entry point.
     Default is the method with the {{:main}} attribute, or else the method named 'Main'.
+/runAllTests:<n> (experimental)
+    0 (default) - Annotates compiled methods with the {{:test}} attribute
+        such that they can be tested using a testing framework
+        in the target language (e.g. xUnit for C#).
+    1 - Emits a main method in the target language that will execute every method
+        in the program with the {{:test}} attribute.
+        Cannot be used if the program already contains a main method.
+        Note that /compile:3 or 4 must be provided as well to actually execute
+        this main method!
 /compileVerbose:<n>
     0 - don't print status of compilation to the console
     1 (default) - print information such as files being written by